pyproject.toml
[project]
name = "fiapx-video-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "celery>=5.2.0",
    "sqlalchemy>=1.4.0",
    "alembic>=1.7.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.5",
    "opencv-python>=4.5.0",
    "redis>=4.0.0",
    "pytest>=6.2.0",
    "pydantic-settings>=2.8.1",
    "psycopg2>=2.9.10",
    "email-validator>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pytest-cov>=6.1.1",
    "httpx>=0.28.1",
]
